+++ wpull/bandwidth.py
@@ -46,17 +46,18 @@ class BandwidthMeter(object):
         '''Return the number of samples collected.'''
         return len(self._samples)
 
-    def feed(self, data_len):
+    def feed(self, data_len, feed_time=None):
         '''Update the bandwidth meter.
 
         Args:
             data_len (int): The number of bytes transfered since the last
                 call to :func:`feed`.
+            feed_time (float): Current time.
         '''
         self._bytes_transferred += data_len
         self._collected_bytes_transferred += data_len
 
-        time_now = time.time()
+        time_now = feed_time or time.time()
         time_diff = time_now - self._last_feed_time
 
         if time_diff < self._sample_min_time:
@@ -89,5 +90,34 @@ class BandwidthMeter(object):
 
         if time_sum:
             return data_len_sum / time_sum
+        else:
+            return 0
+
+
+class BandwidthLimiter(BandwidthMeter):
+    '''Bandwidth rate limit calculator.'''
+    def __init__(self, rate_limit):
+        super().__init__(sample_min_time=0)
+        self._rate_limit = rate_limit
+
+    def sleep_time(self):
+        if not self._samples or not self._rate_limit:
+            return 0
+
+        elapsed_time = 0
+        byte_sum = 0
+
+        for time_diff, data_len in self._samples:
+            elapsed_time += time_diff
+            byte_sum += data_len
+
+        expected_elapsed_time = byte_sum / self._rate_limit
+
+        if expected_elapsed_time > elapsed_time:
+            sleep_time = expected_elapsed_time - elapsed_time
+            if sleep_time < 0.001:
+                return 0
+            else:
+                return sleep_time
         else:
             return 0